+++ Cargo.toml
@@ -2,5 +2,7 @@
 
 members = [
   "substrate-airgapped",
-  "substrate-airgapped-cli",
+  "cli",
+  "metadata",
+  "util"
 ]
