+++ cli/.gitignore
