+++ cli/Cargo.toml
