+++ metadata/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "substrate-airgapped-metadata"
+version = "0.1.0"
+authors = ["emostov <32168567+emostov@users.noreply.github.com>"]
+edition = "2018"
+
+[dependencies]
+# Parity (non-substrate)
+codec = { package = "parity-scale-codec", version = "1.3.5", default-features = false }
+
+# Substrate
+frame-metadata = { version = "12", package = "frame-metadata", default-features = false }
+
+substrate-airgapped = { version = "0.1.0", path = "../substrate-airgapped" }
+
+[dev-dependencies]
+sp-keyring = { version = "2.0.0", package = "sp-keyring", default-features = false }
+hex = { version = "0.4.2" }
+util = { version = "0.1.0", package = "substrate-airgapped-util", path = "../util" }
