+++ substrate-airgapped/Cargo.toml
@@ -4,22 +4,17 @@ version = "0.1.0"
 authors = ["emostov <32168567+emostov@users.noreply.github.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 # Parity (non-substrate)
 codec = { package = "parity-scale-codec", version = "1.3.5", default-features = false }
 
-# Substrate (hopefully most of these can be removed)
-sp-runtime = { version = "2.0.0", package = "sp-runtime" }
-sp-core = { version = "2.0.0" }
-frame-metadata = { version = "12", package = "frame-metadata" }
+# Substrate (hopefully these can be removed in the future)
+sp-runtime = { version = "2.0.0", default-features = false }
+sp-core = { version = "2.0.0", default-features = false }
 
 [dev-dependencies]
 # Currently these are just to get the example going
-sp-keyring = { version = "2.0.0", package = "sp-keyring" }
-sp-version = { version = "2.0.0", package = "sp-version" }
-serde = { version = "1.0.115", features = ["derive"] }
-serde_json = "1.0.57"
+sp-keyring = { version = "2.0.0", package = "sp-keyring", default-features = false }
+sp-version = { version = "2.0.0", package = "sp-version", default-features = false }
 hex = { version = "0.4.2" }
-reqwest = { version = "0.10.10", features = ["json", "blocking"] }
+util = { version = "0.1.0", package = "substrate-airgapped-util", path = "../util" }
