+++ substrate-airgapped/src/frame/mod.rs
@@ -7,7 +7,7 @@ pub mod system;
 ///
 /// When using `decl_module` all arguments for call functions must implement this trait.
 ///
-/// Same as `frame_support::Parameter`. Re-dfined to avoid dependance on frame-support.
+/// Same as `frame_support::Parameter`. Re-defined to avoid dependance on frame-support.
 pub trait Parameter: codec::Codec + Clone + Eq + core::fmt::Debug {}
 impl<T> Parameter for T where T: codec::Codec + Clone + Eq + core::fmt::Debug {}
