+++ substrate-airgapped/src/lib.rs
@@ -2,13 +2,11 @@
 #![warn(missing_docs)]
 
 mod frame;
-mod metadata;
 mod runtimes;
 mod transaction;
 
 pub use crate::{
-	frame::{balances, system},
-	metadata::Metadata,
+	frame::{balances, system, ModuleCall},
 	runtimes::KusamaRuntime,
 	transaction::{
 		tx_from_parts, CallIndex, GenericCall, MortalConfig, Mortality, SignedPayload, Tx,
