+++ util/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "substrate-airgapped-util"
+version = "0.1.0"
+authors = ["emostov <32168567+emostov@users.noreply.github.com>"]
+edition = "2018"
+
+[dependencies]
+# Used in parts of substrate
+serde = { version = "1.0.115", features = ["derive"] }
+serde_json = "1.0.57"
+hex = { version = "0.4.2" }
+
+# Substrate
+sp-core = { version = "2.0.0", default-features = false }
+sp-runtime = { version = "2.0.0", default-features = false }
+
+reqwest = { version = "0.10.10", features = ["json", "blocking"] }
